+++ packages/ricos-editor/web/src/utils/editorUtils.ts
@@ -3,7 +3,12 @@ import { EditorState, ContentState, EditorProps } from 'draft-js';
 import { debounce, pick } from 'lodash';
 import { emptyState, DRAFT_EDITOR_PROPS } from 'ricos-common';
 import { isSSR } from 'wix-rich-content-common';
-import { RicosContent, EditorDataInstance, OnContentChangeFunction } from '../index';
+import {
+  RicosContent,
+  EditorDataInstance,
+  OnContentChangeFunction,
+  ContentStateGetter,
+} from '../index';
 import errorBlocksRemover from './errorBlocksRemover';
 
 /* eslint-disable no-console */
@@ -33,7 +38,7 @@ export function createDataConverter(onContentChange?: OnContentChangeFunction): 
     });
   };
 
-  const getContentState = ({ shouldRemoveErrorBlocks = false } = {}) => {
+  const getContentState: ContentStateGetter = ({ shouldRemoveErrorBlocks = false } = {}) => {
     const currState: ContentState = currEditorState.getCurrentContent();
     if (!isUpdated) {
       currContent = convertToRaw(currState);